Everything related
I  attempted to assimilate the solve website provided with my own web pages. But, I really became confused with the nav vs. the ul in producing my (About Me, Contact, Portfolio) in making them appear correct, and with inter linking them. I think I realize some of the errors, but don't have time to revise without running into issues that can cause it to be worse than this. Will keep plugging to figure it out.  I attempted to add a pic of me in there, but not having any luck with adding the remaining photos.